Header Navigation finish
had a hard time adding background color to the flex items. Solution was to instead of using a unordered list and add the nav texts as "li". I had to create separate div containers for each nav link and its worked.

I was able to color the logo and the div items separately.